refactor(PhoneNumberEditor): Extract component
